Create face_detection_video.cpp
This is a program that uses the OpenCV library to detect faces in a live stream from a webcam or in a video file stored in the local machine. This program shall detect faces in real time and track it. It makes use of pre-trained XML classifiers for the same. These classifiers used in this program have facial features trained in them. Different classifiers can be used to detect different objects. Requirements to run the program: 1) OpenCV should be installed in the local machine. 2) Paths to the classifier XML files should be provided before the program is executed. These XML files can be found in the OpenCV directory "opencv/data/haarcascades". 3) Use 0 in capture.open(0) to play webcam feed. 4) For detection in a local video provide the path to the video.(capture.open("path_to_video").